+++ .env
@@ -13,3 +13,5 @@ PASSWORD_ATLAS = "test12345"
 
 TWILIO_TOKEN = "5b72bdad2c854b9a0bd075358aeb62a2"
 TWILIO_SID = "ACc99fab27c07ad125cc9b9e6e54f863cf"
+
+DB_ATLAS_LINK = "mongodb+srv://pseudobrains:test12345@cluster0.xo0lnsr.mongodb.net/stackDB?retryWrites=true&w=majority"
