+++ controller/logInController.js
@@ -2,7 +2,8 @@ const express = require("express");
 const joi = require("joi");
 const jwt = require("jsonwebtoken");
 const bcrypt = require("bcrypt");
-const { UserModel, AccountDetails } = require("../models/user");
+const { UserModel } = require("../models/user");
+const {AccountDetails} = require("../models/accountDetail");
 const { loginValidation } = require("../models/validation");
 const crypto = require("crypto");
 const { EncryptUserInfo } = require("../util/encrypt");
