+++ models/tokenSchema.js
@@ -1,0 +1,20 @@
+const mongoose = require("mongoose");
+const { Schema, model } = require("mongoose");
+
+//  Schema for reseting password
+const tokenSchema = new Schema({
+    userId: {
+      type: Schema.Types.ObjectId,
+      required: true,
+      ref: "User",
+    },
+    token: { type: String, required: true },
+    createdAt: { type: Date, default: Date.now, expires: 600 },
+   
+  });
+
+  const ResetToken = model("ResetToken", tokenSchema);
+
+
+  
+  module.exports = {ResetToken};
