+++ models/user.js
@@ -52,36 +52,7 @@ const transactionSchema = new Schema({
   },
 });
 
-const accountDetailsSchema = new Schema({
-  userId: {
-    type: Schema.Types.ObjectId,
-    required: true,
-    ref: "User",
-  },
-  balance: {
-    type: mongoose.Decimal128,
-    required: true,
-    default: 0.0,
-  },
-  totalDeposit: {
-    type: mongoose.Decimal128,
-    default: 0.00,
-    trim:true
-  },
 
-  totalWithdraw: {
-    type: mongoose.Decimal128,
-    trim:true,
-    default: 0.00
-  },
-  createdAt: {
-    type: Date,
-    default: () => Date.now(),
-    required:true,
-    immutable:true
-  },
-  updated_at:{ type: Date }
-});
 
 // UserSchema
 
@@ -153,17 +124,7 @@ const UserSchema = new Schema({
 
 //  account Details Schema
 
-//  Schema for reseting password
-const tokenSchema = new Schema({
-  userId: {
-    type: Schema.Types.ObjectId,
-    required: true,
-    ref: "User",
-  },
-  token: { type: String, required: true },
-  createdAt: { type: Date, default: Date.now, expires: 600 },
- 
-});
+
 
 //  Schame middle wares
 
@@ -172,10 +133,6 @@ transactionSchema.pre("save", function (next) {
   next();
 });
 
-accountDetailsSchema.pre("save", function (next) {
-  this.updatedAt = Date.now();
-  next();
-});
 
 UserSchema.pre("save", function (next) {
   this.updatedAt = Date.now();
@@ -184,10 +141,6 @@ UserSchema.pre("save", function (next) {
 
 const UserModel = model("User", UserSchema);
 
-const AccountDetails = model("AccountDetails", accountDetailsSchema);
-
-const ResetToken = model("ResetToken", tokenSchema);
-
-module.exports = { UserModel, AccountDetails, ResetToken };
+module.exports = { UserModel};
