+++ route/reset-password.js
@@ -1,10 +1,7 @@
 const express = require("express");
 const Router = require("express").Router();
 const joi = require("joi");
-const { mailsender } = require("../models/sendMailFun");
-const bcrypt = require("bcrypt");
-const crypto = require("crypto");
-const { ResetToken, UserModel } = require("../models/user");
+const {resetPasswordController} = require("../controller/resetPasswordController")
 
 // Router.get("/reset-password/:id/:token",async (req,res)=>{
 // try {
@@ -26,47 +23,7 @@ const { ResetToken, UserModel } = require("../models/user");
 //     res.status(500).send({message:error,id:user._id})
 // }})
 
-Router.post("/reset-password/:id/:token", async (req, res) => {
-  try {
-    const id = req.params.id.toString();
-    const Ptoken = req.params.token.toString();
-
-    const user = await UserModel.findOne({ _id: id });
-
-    if (!user) return res.status(401).send({ message: "invaild link" });
-    console.log("transaction");
-
-    let token = await ResetToken.findOne({
-      userId: id,
-      token: Ptoken,
-    });
-
-    if (!token) return res.status(400).send({ message: "invaild link" });
-
-    const PasswordSchema = joi.object({
-      password: joi.string().min(6).required(),
-    });
-
-    const { error } = PasswordSchema.validate(req.body);
-
-    console.log(token, error);
-
-    if (error)
-      return res.status(400).send({ message: error.details[0].message });
-
-    const salt = await bcrypt.genSalt(10);
-
-    //         const hashPassword = await bcrypt.hash(req.body.password,salt);
-
-    await UserModel.findOneAndUpdate({ _id: id }, { password: hashPassword });
-
-    //          await ResetToken.findOneAndDelete({token: token.tokens})
-
-    res.status(200).send({ message: "success" });
-  } catch (error) {
-    res.status(500).send({ message: error });
-  }
-});
+Router.post("/reset-password/:id/:token",resetPasswordController);
 
     module.exports={
     ResetPasswordRoute: Router
